moving get_train_args to config
